+++ doc/ChangeLog.md
@@ -17,6 +17,7 @@ All notable changes to the project are documented in this file.
    RPC and add boot order to operational datastore.
  - SSH Server is now configurable, issue #441
    SSH Server and NETCONF Server now uses the same SSH hostkey in factory-config
+ - Add support for GRE/GRETAP tunnels
 
 ### Fixes